refactor(dev): keep lease admission below module ceilings
Docs-Impact-Decision: Host-resource behavior is documented in architecture, operations, and the contributor guide; enrolling its test in the existing policy profile has no additional user-facing effect.
