+++ apps/web/lib/docs/doc-impact.generated.json
@@ -841,6 +841,14 @@
       "docs/architecture/section-navigation.md",
       "docs/maintenance/staleness-report.md"
     ],
+    "apps/web/lib/nonprod/host-resource-policy.ts": [
+      "docs/architecture/dpf-patterns.md",
+      "docs/user-guide/contributing/agent-dev-environments.md"
+    ],
+    "apps/web/lib/nonprod/host-resource-profiles.json": [
+      "docs/architecture/dpf-patterns.md",
+      "docs/user-guide/contributing/agent-dev-environments.md"
+    ],
     "apps/web/lib/operate/db-continuity.ts": [
       "docs/architecture/delivery-surfaces-runbook.md"
     ],
@@ -1262,7 +1270,8 @@
     ],
     "packages/db/prisma/schema/build-delivery.prisma": [
       "docs/architecture/2026-06-09-long-running-agentic-process-architecture.md",
-      "docs/architecture/dpf-patterns.md"
+      "docs/architecture/dpf-patterns.md",
+      "docs/user-guide/contributing/agent-dev-environments.md"
     ],
     "packages/db/prisma/schema/resource-scheduling.prisma": [
       "docs/architecture/data-model-stewardship-runbook.md"
@@ -1684,6 +1693,11 @@
     "scripts/hooks/run-hook.mjs": [
       "docs/install/platform-support-watchlist.md",
       "docs/operations/session-transcript-recovery.md"
+    ],
+    "scripts/host-resource-runner.mjs": [
+      "docs/architecture/dpf-patterns.md",
+      "docs/operations/local-ci-sandbox-slots.md",
+      "docs/user-guide/contributing/agent-dev-environments.md"
     ],
     "scripts/install-worktree-janitor-schedule.sh": [
       "docs/install/worktree-janitor-schedule.md"
@@ -2376,8 +2390,11 @@
       "scripts/sbom/scan-dependencies.mjs"
     ],
     "docs/architecture/dpf-patterns.md": [
+      "apps/web/lib/nonprod/host-resource-policy.ts",
+      "apps/web/lib/nonprod/host-resource-profiles.json",
       "packages/db/prisma/schema/build-delivery.prisma",
-      "packages/db/prisma/schema/work-coordination.prisma"
+      "packages/db/prisma/schema/work-coordination.prisma",
+      "scripts/host-resource-runner.mjs"
     ],
     "docs/architecture/enforced-ci-gates.md": [
       "apps/web/lib/docs-route-map.ts",
@@ -2699,6 +2716,7 @@
     ],
     "docs/operations/local-ci-sandbox-slots.md": [
       "scripts/check-host-port-range.test.mjs",
+      "scripts/host-resource-runner.mjs",
       "scripts/lib/local-ci-slot-manifest.mjs",
       "scripts/local-ci-pilot-report.mjs"
     ],
@@ -2994,8 +3012,12 @@
       "apps/web/lib/build/build-orchestrator.ts"
     ],
     "docs/user-guide/contributing/agent-dev-environments.md": [
+      "apps/web/lib/nonprod/host-resource-policy.ts",
+      "apps/web/lib/nonprod/host-resource-profiles.json",
+      "packages/db/prisma/schema/build-delivery.prisma",
       "scripts/dpf-bootstrap-agent-toolchain.ps1",
-      "scripts/dpf-bootstrap-agent-toolchain.sh"
+      "scripts/dpf-bootstrap-agent-toolchain.sh",
+      "scripts/host-resource-runner.mjs"
     ],
     "docs/user-guide/contributing/dev-container.md": [
       "packages/db/scripts/dev-preview-migrate-deploy.mjs"
