+++ docs/superpowers/plans/2026-08-24-resilient-gate-flow-control.md
@@ -99,7 +99,7 @@ Increase delivery throughput without reducing evidence quality by moving wait st
 the existing `NonProductionEnvironmentLease` at high confidence. The host lane
 therefore does not create a new broker/table or encode its contract into a
 free-form purpose string. `resourceClass`, `expectedMemoryBytes`, and
-`ownerProcessId` extend the existing durable authority; the in-process
+`ownerPid` extend the existing durable authority; the in-process
 `ResourceLane` remains its portal adapter. `ownerProcessIdentity` accompanies
 the PID so liveness cannot be fooled by OS PID reuse.
